requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
google-generativeai==0.3.0
python-dotenv==1.0.0
datetime==5.1
uuid==1.30
cryptography==41.0.7
gunicorn==20.1.0

pymysql==1.1.0
psycopg2-binary==2.9.9